refactor(search): split line construction helpers
Move native and OCR search line reconstruction plus bbox slicing out of the search execution module so matcher orchestration stays separate from text-line reconstruction.
